move dig logic to dig from pathable
