+++ plugins/CMakeLists.txt
@@ -106,7 +106,7 @@ if(BUILD_SUPPORTED)
     dfhack_plugin(cxxrandom cxxrandom.cpp LINK_LIBRARIES lua)
     dfhack_plugin(deramp deramp.cpp)
     dfhack_plugin(debug debug.cpp LINK_LIBRARIES jsoncpp_static)
-    dfhack_plugin(dig dig.cpp)
+    dfhack_plugin(dig dig.cpp LINK_LIBRARIES lua)
     dfhack_plugin(dig-now dig-now.cpp LINK_LIBRARIES lua)
     #dfhack_plugin(digFlood digFlood.cpp)
     #add_subdirectory(diggingInvaders)
